Review of DB schema and models.py update
We seem to have only one start/end pair for the whole `Configuration` but each `ExpectedMeasure` didn't have a way of defining an interval for the day. Added a start/end hour/minute for each `ExpectedMeasure` as to define the from 0000 until 2359 schedule for a day. This is reflected in the `database_design.md` and in the
`models.py`.

A `Configuration` can have a start time from when it is applied, but not necessarily an end_time, this property is now nullable.

Added some back_populate properties to `ExpectedMeasure` and `Configuration`.

Re-sourced some imports as these were unnecessarily tying the configuration in with the models and modelschemas.